backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
mysqlclient==2.2.4
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
